test(plugins): separate inspect command coverage
